TNL-5466 tests for changing grade policy and moving progress page tests into separate file
