+++ common/test/acceptance/pages/lms/problem.py
@@ -162,7 +162,7 @@ class ProblemPage(PageObject):
         """
         Click the choice input(radio, checkbox or option) where value matches `choice_value` in choice group.
         """
-        self.q(css='div.problem .choicegroup input[value="' + choice_value + '"]').click()
+        self.q(css='div.problem .choicegroup input[value="' + choice_value + '"]').first.click()
         self.wait_for_ajax()
 
     def is_correct(self):
